Change the loop index from i=0 to i=n
Per the suggestion of the section 5.3 C Looping Structures of the
ARM System Developer's Guide, loop index with a fixed number of
interations should be started from i=n (decrement the loop index
for each iteration). This approch can save one instruction comparing
to the iteration with incrementing the loop index.

BTW, this commit also eliminates the unused register 'r3'.